--- tgui/packages/tgui/interfaces/PreferencesMenu/preferences/features/character_preferences/types.ts
@@ -1,8 +1,0 @@
-type BackgroundInformationData = {
-  slot: number,
-  flavor_text: string,
-  med_record: string,
-  gen_record: string,
-  sec_record: string,
-  exploit_record: string,
-}
